pyproject.toml
[tool.poetry]
name = "Soccer Manager"
version = "0.1.0"
description = "Soccer online manager game API"
authors = ["marius.crisan <marius.crisan.tech@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
faker = "^22.0.0"
django-password-validators = "^1.7.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
